Huge commit: this is because this is not the test repository.
As this commit is huge, we'll take some time to explain some of the
changes. The general idea is that we're going towards more modularity.

First, we have some simple renaming: control promises files all now
start with "control_", while bundle promises files start with "bundle_".

We removed the main_* files, creating in the same time bundle files for
every bundle previously contained in the main_files.

We added some templating, files to deploy being stored in deployfiles/,
while the templating process is described in bundle_deploy_files.cf.